Move `info` and `cumulative_info` to `Proof` trait
